fuzzgen: Move `FunctionWithIsa` to icache fuzzer
